\name{QUALDIAG}
\alias{QUALDIAG}
\title{ Produces Diagnostic Graphics }
\description{
Internal function, produces graphics and statistics outputs. 
Please, consult additionnal documentation for computation details.
}
\usage{
QUALDIAG(...)
}

\value{
A list (diag.values) with:
  \itemize{ 
    \item{bands.widths }{Bins Widths (bp)}
    \item{bands.pos}{Bins Sizes (bp)}
    \item{Nbins}{Number of bins}
    \item{homoplasy}{Mean Homoplasy}
  }
}
\references{ ~put references to the literature/web site here ~ }
\author{ Nils Arrigo (nils.arrigo@unine.ch), University Of Neuchatel, Laboratory of Evolutive Botany, 2009 Neuchatel, Switzerland 
         http://www.unine.ch/bota/EBOLAB/}
