\name{RawGeno 2.0-1 package}
\alias{RawGeno-package}
\alias{RawGeno}
\docType{package}
\title{
RawGeno
}
\description{
Processes raw genotyping outputs in order to facilitate data scoring.
}
\details{
\tabular{ll}{
Package: \tab RawGeno\cr
Type: \tab Package\cr
Version: \tab 1.07\cr
Date: \tab 2012-03-20\cr
License: \tab LGPL\cr
}
The package currently processes outputs of GeneScan (produced by ABI Genscan 3.7). 
Once exported in a *.txt file, chromatograms of several samples can be merged. 
The final output is a binary matrix (samples x peaks). More functions will be added for agarose or acrylamide-gel pictures handling.
}
\author{
Nils Arrigo (nils.arrigo@unine.ch), University Of Neuchatel, Laboratory of Evolutive Botany, 2009 Neuchatel, Switzerland 
         http://www.unine.ch/bota/EBOLAB/

Maintainer: nils.arrigo@unine.ch
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ Import  }
\keyword{ Genescan }
\keyword{ Genotype }
\keyword{ AFLP }
\keyword{ Dominant }
\keyword{ Binary }
\keyword{ Genotyping }


\seealso{
\code{\link[<pkg>:<pkg>-package]{<pkg>}}
}
