\name{SAVE.GUI}
\alias{SAVE.GUI}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Exporting Functions GUI }
\description{
Supervises the exporting functions.
}
\usage{
SAVE.GUI(...)
}
\author{ Nils Arrigo (nils.arrigo@unine.ch), University Of Neuchatel, Laboratory of Evolutive Botany, 2009 Neuchatel, Switzerland 
         http://www.unine.ch/bota/EBOLAB/}
\examples{
## Runs only if the scoring phase has been achieved.
#SAVE.GUI()
}
