\name{Spagedi.Ind}
\alias{Spagedi.Ind}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Exporting function to Spagedi - Individual level }
\description{
Function to make an input file for Spagedi, at the individuals level
}
\usage{
Spagedi.Ind(mat, path)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{mat}{ A binary matrix, individuals in rows and markers in column }
  \item{path}{ File path to produce the export }
}
\references{ Ehrich, D. 2006. AFLPdat: a collection of R functions for convenient handling of AFLP
data. Mol. Ecol. Notes, in press. }
\author{ Dorothée Ehrich }